backend/requirements.txt
fastapi>=0.112
uvicorn[standard]>=0.30
pydantic[email]>=2
prisma>=0.15.0
bcrypt>=4.0.0
resend>=0.8.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
requests>=2.31.0
pyotp>=2.9.0
httpx>=0.24.0
python-dotenv>=1.0.0